refactor(cli): split semantic completion error mapper
