+++ crates/ctx-cli/src/dispatch.rs
@@ -43,6 +43,7 @@ use ctx_app_config::{AppConfig, DeprecatedControls};
 
 mod finalization;
 mod parse;
+mod semantic_completion_error;
 #[cfg(test)]
 mod test_support;
 
@@ -584,7 +585,7 @@ fn render_command_result_error(
                 write_machine_error(
                     search_operation,
                     ui,
-                    &serde_json::to_string(&semantic_completion_error_json(error))?,
+                    &serde_json::to_string(&semantic_completion_error::structured(error))?,
                 )?;
                 Some(RenderedJsonError.into())
             } else {
@@ -599,46 +600,6 @@ fn render_command_result_error(
         None
     };
     Ok(rendered_error)
-}
-
-fn semantic_completion_error_json(
-    error: &ctx_daemon_cli::SemanticCompletionError,
-) -> serde_json::Value {
-    let detail = error.to_string();
-    let mut structured = serde_json::json!({
-        "error": detail,
-        "error_code": "semantic_completion_failed",
-        "reason": error.code(),
-        "generation_id": error.generation_id(),
-        "core_published": true,
-        "retryable": error.retryable(),
-        "detail": detail,
-    });
-    let fields = structured
-        .as_object_mut()
-        .expect("semantic completion error JSON must be an object");
-    match error {
-        ctx_daemon_cli::SemanticCompletionError::CoreSuperseded {
-            active_generation_id,
-            ..
-        } => {
-            fields.insert(
-                "active_generation_id".to_owned(),
-                serde_json::Value::String(active_generation_id.clone()),
-            );
-        }
-        ctx_daemon_cli::SemanticCompletionError::DaemonJobFailed {
-            failure_class: Some(failure_class),
-            ..
-        } => {
-            fields.insert(
-                "failure_class".to_owned(),
-                serde_json::Value::String(failure_class.clone()),
-            );
-        }
-        _ => {}
-    }
-    structured
 }
 
 fn write_machine_error(search_operation: bool, ui: &mut Ui, message: &str) -> Result<()> {
